Unify Errors into one file
