+++ src/structures/errors.js
@@ -39,4 +39,25 @@ class Ratelimit extends Error {
 	}
 }
 
-module.exports = Ratelimit;
+/**
+ * FetchError
+ * @extends {Error}
+ */
+class FetchError extends Error {
+	/**
+     * @param {Response} i
+	 * @param {string} message
+     */
+	constructor(i, message) {
+		super(`${i.status} ${message}`);
+		Object.defineProperty(this, 'name', { value: 'FetchError' });
+		if (Error.captureStackTrace) Error.captureStackTrace(this, FetchError);
+	}
+
+	toString() {
+		return this.message;
+	}
+}
+
+exports.FetchError = FetchError;
+exports.Ratelimit = Ratelimit;
